+++ tests/test_config.py
@@ -1,12 +1,12 @@
 # coding=utf-8
 
-from xpaw.config import BaseConfig, Config, KNOWN_SETTINGS
+from xpaw.config import Config
 
 
-class TestBaseConfig:
+class TestConfig:
     def test_get(self):
         d = {'key': 'value'}
-        config = BaseConfig(d)
+        config = Config(d)
         assert len(config) == 1
         for k in config:
             assert k == 'key' and config[k] == 'value'
@@ -23,7 +23,7 @@ class TestBaseConfig:
         d = {'bool_true': 'true', 'bool_True': 'True', 'bool_false': 'false', 'bool_False': 'False',
              'bool_int1': '1', 'bool_int0': '0', 'bool_none': '...',
              'true': True, 'false': False}
-        config = BaseConfig(d)
+        config = Config(d)
         assert config.getbool('true') is True
         assert config.getbool('false') is False
         assert config.getbool('bool_true') is True
@@ -42,7 +42,7 @@ class TestBaseConfig:
 
     def test_get_int(self):
         d = {'int_1': 1, 'int_str_1': '1', 'int_none': '...'}
-        config = BaseConfig(d)
+        config = Config(d)
         assert config.getint('int_1') == 1
         assert config.getint('int_str_1') == 1
         assert config.getint('int_none') is None
@@ -53,7 +53,7 @@ class TestBaseConfig:
 
     def test_get_float(self):
         d = {'float_1.1': 1.1, 'float_str_1.1': '1.1', 'float_none': '...'}
-        config = BaseConfig(d)
+        config = Config(d)
         assert config.getfloat('float_1.1') == 1.1
         assert config.getfloat('float_str_1.1') == 1.1
         assert config.getfloat('float_none') is None
@@ -64,7 +64,7 @@ class TestBaseConfig:
 
     def test_get_list(self):
         d = {'list': [1, 2], 'tuple': (1, 2), 'single': 1, 'list_str': '1,2'}
-        config = BaseConfig(d)
+        config = Config(d)
         assert config.getlist('list') == [1, 2]
         assert config.getlist('tuple') == [1, 2]
         assert config.getlist('single') == [1]
@@ -74,7 +74,7 @@ class TestBaseConfig:
         assert config.getlist('no_such_list', [1]) == [1]
 
     def test_set(self):
-        config = BaseConfig()
+        config = Config()
         config.set('key', 'value')
         assert len(config) == 1 and config['key'] == 'value'
 
@@ -85,7 +85,7 @@ class TestBaseConfig:
         assert len(config) == 2 and config['key2'] == 'value'
 
     def test_set_item(self):
-        config = BaseConfig()
+        config = Config()
         config['key'] = 'value'
         assert len(config) == 1 and config['key'] == 'value'
 
@@ -96,16 +96,16 @@ class TestBaseConfig:
         assert len(config) == 2 and config['key2'] == 'value'
 
     def test_update(self):
-        config = BaseConfig()
+        config = Config()
         config.update({'key': 'value'})
         assert len(config) == 1 and config['key'] == 'value'
 
         config.update({'key': 'value2', 'key2': 'value'})
         assert len(config) == 2 and config['key'] == 'value2' and config['key2'] == 'value'
 
-    def test_update_by_base_config(self):
-        c1 = BaseConfig({'k1': 'c1_k1'})
-        c2 = BaseConfig({'k1': 'c2_k1', 'k2': 'c2_k2'}, k3='c2_k3')
+    def test_update_by_config(self):
+        c1 = Config({'k1': 'c1_k1'})
+        c2 = Config({'k1': 'c2_k1', 'k2': 'c2_k2'}, k3='c2_k3')
         c1.update(c2)
         assert len(c1) == 3
         assert c1['k1'] == 'c2_k1'
@@ -113,13 +113,13 @@ class TestBaseConfig:
         assert c1['k3'] == 'c2_k3'
 
     def test_copy(self):
-        c1 = BaseConfig({'dict': {'k': 'v'}})
+        c1 = Config({'dict': {'k': 'v'}})
         c2 = c1.copy()
         c1['dict']['k'] = 'vv'
         assert c2['dict']['k'] == 'v'
 
     def test_delete(self):
-        c = BaseConfig({'k1': 'v1', 'k2': 'v2', 'k3': 'v3'})
+        c = Config({'k1': 'v1', 'k2': 'v2', 'k3': 'v3'})
         c.delete('k2')
         assert len(c) == 2 and 'k2' not in c
         c.delete('k1')
@@ -128,15 +128,8 @@ class TestBaseConfig:
         assert len(c) == 0 and 'k3' not in c
 
     def test_setdefault(self):
-        c = BaseConfig()
+        c = Config()
         c.setdefault('k1', 'v1')
         assert c['k1'] == 'v1'
         c.setdefault('k1', 'v2')
         assert c['k1'] == 'v1'
-
-
-def test_config():
-    c = Config()
-    assert len(c) > 0
-    for k in c:
-        assert k in KNOWN_SETTINGS and c.get(k) == KNOWN_SETTINGS[k].value