+++ tests/test_extension.py
@@ -3,7 +3,7 @@
 import pytest
 
 from xpaw.extension import ExtensionManager
-from xpaw.config import Config
+from xpaw.config import Config, DEFAULT_CONFIG
 from xpaw.eventbus import EventBus
 from xpaw import events
 
@@ -28,7 +28,7 @@ class DummyExtension:
 class Cluster:
     def __init__(self, **kwargs):
         self.event_bus = EventBus()
-        self.config = Config(kwargs)
+        self.config = Config(DEFAULT_CONFIG, **kwargs)
 
 
 @pytest.mark.asyncio